src/main/trials/requirements.txt
python-dotenv
openai
pyyaml
google-generativeai
gradio
pydantic
pypdf
openai-agents
sib-api-v3-sdk